requirement.txt
selenium 
bs4
requests
pandas
web-driver-manager
openpyxl
googlesearch-python